Improve complication editor UI
